Separated googleDriveSync into modules
