+++ client/src/app/components/pageManager/cloudProviderPage/providerItem/useProviderItem.ts
@@ -4,7 +4,7 @@ import OneDriveLogo from '@assets/images/onedrive.png';
 import VaultLogo from '@assets/images/vault.png';
 import { ConfigDispatch } from '@types';
 // import { getGoogleAuth } from '@shared/api/getGoogleAuth';
-import { getGoogleAuth } from '@shared/api/googleDriveSync';
+import { getGoogleAuth } from '@shared/Api/googleDrive/setup';
 import { useHistory } from 'react-router-dom';
 
 type SupportedProvider = 'Google Drive' | 'OneDrive' | 'PrivaNote Vault';
