+++ client/src/app/hooks/useIpcListeners.ts
@@ -1,6 +1,5 @@
 import { verifyUser } from '@shared/Api/verifyUser';
 import { getFileName, parseCodeFromUrl } from '@shared/utils';
-import { exportNotebookStructure } from '@shared/utils/synchronization/exportNotebookStructure';
 import { ipcRenderer } from 'electron';
 import { useEffect } from 'react';
 import { useHistory } from 'react-router';
@@ -46,9 +45,6 @@ export function useIpcListeners() {
 					type: 'LOAD',
 					payload: location
 				});
-
-				exportNotebookStructure(location);
-				ipcRenderer.send('notebookRootLocation', location);
 			}
 		);
