+++ client/src/shared/utils/synchronization/exportNotebookStructure.ts
@@ -2,8 +2,16 @@ import { saveFile } from '../fileSystem/saveFile';
 import { FileItem } from '../../types';
 import { createNotebookStructure } from './createNotebookStructure';
 
-export const exportNotebookStructure = (path: string) => {
-	const notebookStructure = createNotebookStructure(path);
+export const exportNotebookStructure = (
+	path: string,
+	structure: any = undefined
+) => {
+	let notebookStructure: any;
+	if (structure) {
+		notebookStructure = structure;
+	} else {
+		notebookStructure = createNotebookStructure(path);
+	}
 
 	const name = 'notebookStructure.json';
 	const exportFile: FileItem = {
